Rename folders to lib/core
