+++ lib/core/encoder.ts
