+++ lib/core/encoders/error.ts
