+++ lib/core/error.ts
