+++ lib/core/opts.ts
