+++ lib/core/promise.ts
