+++ tsconfig.json
@@ -101,7 +101,7 @@
       "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
     },
     "include": [
-      "src/**/*",
+      "lib/**/*",
       "test/**/*"
     ]
   }
